Bot works in dm now, music commands don't work, and you cant remove other's reactions in dm's so meep also doesn't work will try to make this work later

+++ src/commands/alias.js
@@ -6,22 +6,22 @@ const {MessageEmbed} = require('discord.js');
 const {getGetString} = require('../utils/strings');
 
 async function alias(msg, msgArgs) {
-	const getString = getGetString(msg.guild.id);
+	const getString = getGetString((msg.guild || msg.author).id);
 	const newCommand = msgArgs[0];
 	if (msgArgs.length > 1) {
 		if (commands[newCommand] || newCommand === 'help') {
 			await msg.channel.send(getString('alias_override'));
-		} else if (newCommand.startsWith(database.get(msg.guild.id, 'settings.prefix'))) {
+		} else if (newCommand.startsWith(database.get((msg.guild || msg.author).id, 'settings.prefix'))) {
 			await msg.channel.send(getString('alias_prefix'));
 		} else {
 			let command = msgArgs[1];
-			if (!commands[command] && !database.has(msg.guild.id, `aliases.${command}`) && command !== 'help') {
+			if (!commands[command] && !database.has((msg.guild || msg.author).id, `aliases.${command}`) && command !== 'help') {
 				return msg.channel.send(getString('alias_not_command').format(command));
 			}
 			const args = msgArgs.slice(2);
 
-			const existed = database.has(msg.guild.id, `aliases.${newCommand}`);
-			database.set(msg.guild.id, {
+			const existed = database.has((msg.guild || msg.author).id, `aliases.${newCommand}`);
+			database.set((msg.guild || msg.author).id, {
 				command,
 				args
 			}, `aliases.${newCommand}`);
@@ -34,11 +34,11 @@ async function alias(msg, msgArgs) {
 
 const aliasrm = {
 	'function': async (msg, [alias]) => {
-		const getString = getGetString(msg.guild.id);
+		const getString = getGetString((msg.guild || msg.author).id);
 		if (alias) {
-			const exists = database.has(msg.guild.id, `aliases.${alias}`);
+			const exists = database.has((msg.guild || msg.author).id, `aliases.${alias}`);
 			if (exists) {
-				database.delete(msg.guild.id, `aliases.${alias}`);
+				database.delete((msg.guild || msg.author).id, `aliases.${alias}`);
 				await msg.channel.send(getString('aliasrm_removed').format(alias));
 			} else {
 				await msg.channel.send(getString('aliasrm_find').format(alias));
@@ -52,13 +52,13 @@ const aliasrm = {
 
 const aliases = {
 	'function': async (msg, [arg1]) => {
-		const getString = getGetString(msg.guild.id);
+		const getString = getGetString((msg.guild || msg.author).id);
 
 		if (arg1 === 'reset') {
-			database.set(msg.guild.id, defaultAliases, 'aliases');
+			database.set((msg.guild || msg.author).id, defaultAliases, 'aliases');
 			await msg.channel.send(getString('aliases_reset'));
 		} else {
-			const aliases = database.get(msg.guild.id, 'aliases');
+			const aliases = database.get((msg.guild || msg.author).id, 'aliases');
 			const values = [];
 
 			const entries = Object.entries(aliases);
